Moved Coordinate Class to upper level
