+++ src/main/java/tictactoe/Board.java
@@ -1,7 +1,6 @@
 package tictactoe;
 
 import java.util.LinkedHashMap;
-import java.util.Objects;
 
 public class Board {
 
@@ -91,37 +90,5 @@ public class Board {
 		return (expected == char1) && (expected == char2) && (expected == char3);
 	}
 
-
-	public static class Coordinate {
-
-		private final int x;
-		private final int y;
-
-		public Coordinate(int x, int y) {
-			this.x = x;
-			this.y = y;
-		}
-
-		public int getX() {
-			return x;
-		}
-
-		public int getY() {
-			return y;
-		}
-
-		@Override
-		public boolean equals(Object o) {
-			if (this == o) return true;
-			if (!(o instanceof Coordinate)) return false;
-			Coordinate that = (Coordinate) o;
-			return getX() == that.getX() && getY() == that.getY();
-		}
-
-		@Override
-		public int hashCode() {
-			return Objects.hash(getX(), getY());
-		}
-	}
 
 }
